fix errors, deleted unecessary methods
