+++ agricolae/app/Http/Controllers/Api/ProductApi.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Http\Controllers\Api;
+
+use App\Http\Controllers\Controller;
+use App\Product;
+use Illuminate\Http\Request;
+
+class ProductApi extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */ 
+    public function index()
+    {
+        return Product::paginate();
+    }
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Product  $product
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        return Product::findOrFail($id);
+    }
+}
