--- agricolae/app/Http/Controllers/Api/ProductController.php
@@ -1,91 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Api;
-
-use App\Http\Controllers\Controller;
-use App\Product;
-use Illuminate\Http\Request;
-
-class ProductController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */ 
-    public function index()
-    {
-        return response()->json(Product::paginate());
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        $request->validate(Product::validateRules());
-        $name = "";
-
-        if ($request->hasFile('image'))
-        {
-            $file = $request->file('image');
-            $name = time().$file->getClientOriginalName();
-            $file->move(public_path().'/images/products_images/', $name);
-        }
-
-        $product = new Product;
-        $product->user_id = 1;//$user->getId();
-        $product->name = $request["name"];
-        $product->description = $request["description"];
-        $product->category = $request["category"];
-        $product->price = $request["price"];
-        $product->units = $request["units"];
-        $product->image = $name;
-        $product->save();
-
-        return response()->json($product, 201);
-
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Product  $product
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Product $product)
-    {
-        return response()->json($product, 200);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Product  $product
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Product $product)
-    {
-        $product->update($request->all());
-
-        return response()->json($product, 200);
-
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Product  $product
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Product $product)
-    {
-        $product->delete();
-
-        return response()->json(null, 204);
-    }
-}
