+++ agricolae/routes/api.php
@@ -18,4 +18,4 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::apiResource('products', 'Api\ProductController')->middleware('auth:api');
+Route::apiResource('products', 'Api\ProductApi');
